requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.26
psycopg2-binary==2.9.9
alembic==1.12.0
python-jose==3.3.0
passlib[argon2]==1.7.4
pydantic==1.10.11
pytest==7.4.0
httpx==0.25.1
python-dotenv==1.0.0
python-jose==3.3.0
pytz-2025.2

